# backend/app/Application/DTOs/Product_DTO.php
<?php

namespace App\Application\DTOs;

use App\Domain\Entities\Product;

/**
 * @OA\Schema(
 *     schema="Product_DTO",
 *     type="object",
 *     required={"Name","Category"},
 *     @OA\Property(property="ProductID", type="integer", nullable=true),
 *     @OA\Property(property="Name", type="string"),
 *     @OA\Property(property="Description", type="string", nullable=true),
 *     @OA\Property(property="Price", type="number", format="float", nullable=true),
 *     @OA\Property(property="IsFavorite", type="boolean"),
 *     @OA\Property(property="Category", type="string"),
 * )
 */
class Product_DTO
{
    public ?int $ProductID = null;
    public string $Name;
    public ?string $Description = null;
    public ?float $Price = null;
    public bool $IsFavorite = false;
    public string $Category;

    public function __construct(?Product $product = null)
    {
        if ($product) {
            $this->ProductID = $product->ProductID;
            $this->Name = $product->Name;
            $this->Description = $product->Description;
            $this->Price = $product->Price;
            $this->IsFavorite = $product->IsFavorite;
            $this->Category = $product->Category;
        }
    }
}