# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "This is the API documentation for my app",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Default Server"
        }
    ],
    "paths": {
...
        "/api/product/{ProductID}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get a product by ProductID",
                "operationId": "3e50d492f7b59e830d6a06e2cb0de449",
                "parameters": [
                    {
                        "name": "ProductID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product_DTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/api/product/create": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create a new product",
                "operationId": "fa18b92abe1bc7e2b5c6b50b717c960a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product_DTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created product",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product_DTO"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/product/update/{ProductID}": {
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Update a product",
                "operationId": "83c983e4dbf50d21817349b633363f80",
                "parameters": [
                    {
                        "name": "ProductID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product_DTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated product",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product_DTO"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/api/product/delete/{ProductID}": {
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete a product",
                "operationId": "596bc4d5083abf4627a6f0fe57c0b66b",
                "parameters": [
                    {
                        "name": "ProductID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "400": {
                        "description": "Deletion failed"
                    }
                }
            }
        },
        "/api/product-market/assign/{ProductID}": {
            "post": {
                "tags": [
                    "ProductMarket"
                ],
                "summary": "Attach a product to selected markets dynamically",
                "operationId": "b216ae1ed0c813efc200df4b74167892",
                "parameters": [
                    {
                        "name": "ProductID",
                        "in": "path",
                        "description": "ID of the product to assign markets to",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "market_ids": {
                                        "description": "Array of market IDs selected by the user",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product successfully assigned to markets",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product_DTO": {
                "required": [
                    "Name",
                    "Category"
                ],
                "properties": {
                    "ProductID": {
                        "type": "integer",
                        "nullable": true
                    },
                    "Name": {
                        "type": "string"
                    },
                    "Description": {
                        "type": "string",
                        "nullable": true
                    },
                    "Price": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "IsFavorite": {
                        "type": "boolean"
                    },
                    "Category": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}